requirements.txt
PyTumblr==0.0.6
PyYAML==3.11
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
facebook-sdk==0.4.0
httplib2==0.9
httpretty==0.8.4
matplotlib==1.4.3
mock==1.0.1
nose==1.3.4
numpy==1.9.1
oauth2==1.5.211
oauthlib==0.7.2
pyparsing==2.0.3
python-dateutil==2.4.0
python-instagram==1.3.0
pytz==2014.10
requests==2.4.3
requests-oauthlib==0.4.1
scikit-learn==0.15.2
scipy==0.15.1
simplejson==3.6.5
six==1.7.3
tornado==4.1
tweepy==3.2.0
urllib3==1.10.1
wsgiref==0.1.2
